WIP ranking + simulator changes

+++ src/main/java/link/locutus/discord/commands/manager/v2/impl/pw/ranking/RankingSectionSpec.java
@@ -1,5 +1,6 @@
 package link.locutus.discord.commands.manager.v2.impl.pw.ranking;
 
+import java.math.BigDecimal;
 import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -8,7 +9,7 @@ import java.util.Objects;
 public record RankingSectionSpec(
         RankingSectionKind sectionKind,
         RankingSortDirection sortDirection,
-        Map<Integer, Double> values
+    Map<Integer, BigDecimal> values
 ) {
     public RankingSectionSpec {
         sectionKind = Objects.requireNonNull(sectionKind, "sectionKind");
